HEEDLS-453 Refactor tests and further formatting changes
