+++ .editorconfig
@@ -11,7 +11,7 @@ resharper_csharp_wrap_after_declaration_lpar = true
 resharper_csharp_wrap_after_invocation_lpar = true
 resharper_csharp_wrap_arguments_style = chop_if_long
 resharper_csharp_wrap_before_declaration_lpar = false
-resharper_csharp_wrap_before_invocation_lpar = true
+resharper_csharp_wrap_before_invocation_lpar = false
 resharper_csharp_wrap_lines = true
 resharper_csharp_wrap_parameters_style = chop_if_long
 resharper_html_space_before_self_closing = true
