+++ DigitalLearningSolutions.Web/appsettings.Development.json
@@ -1,6 +1,6 @@
 {
   "ConnectionStrings": {
-    "DefaultConnection": "Data Source=localhost;Initial Catalog=mbdbx101_test;Integrated Security=True;"
+    "DefaultConnection": "Data Source=localhost;Initial Catalog=mbdbx101;Integrated Security=True;"
   },
   "CurrentSystemBaseUrl": "https://localhost:44367",
   "AppRootPath": "https://localhost:44363",
